requirements.txt
openai>=0.27.0
faiss-cpu>=1.7.3
numpy>=1.23.0
pandas>=1.5.0
python-dotenv>=1.0.0
langchain-community>=0.2.0
pytest>=7.0.0
